refactor structure + svpg_a2c_mono
